pack packets into large buffers when GSO is available

+++ sys_conn_no_gso.go
@@ -4,5 +4,5 @@ package quic
 
 import "syscall"
 
-func maybeSetGSO(_ syscall.RawConn) bool             { return false }
-func appendUDPSegmentSizeMsg(_ []byte, _ int) []byte { return nil }
+func maybeSetGSO(_ syscall.RawConn) bool                { return false }
+func appendUDPSegmentSizeMsg(_ []byte, _ uint16) []byte { return nil }
